UPSTREAM: dma-buf/sync_file: hold reference to fence when creating sync_file
fence referencing was out of balance. It was not taking any ref to the
fence at creating time, but it was putting a reference when freeing the
sync file.

This patch fixes the balancing issue by getting a reference for the fence
when creating the sync_file.

BUG=chromium:911825
TEST=boot caroline, chell, run login_LoginSuccess, graphics_SanAngeles, cheets_StartAndroid

Signed-off-by: Gustavo Padovan <gustavo.padovan@collabora.co.uk>
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Link: http://patchwork.freedesktop.org/patch/msgid/1476899313-22241-1-git-send-email-gustavo@padovan.org
(cherry picked from commit 30cd85dd6edc86ea8d8589efb813f1fad41ef233)
Signed-off-by: Dominik Behr <dbehr@chromium.org>

Change-Id: Ia6df338ac4134a3f892ea019581ce79f02667dac