UPSTREAM: drm/i915: Remove superfluous wait_for_error() from throttle-ioctl
The throttle-ioctl never touches the struct_mutex. It does, however, as
part of its ABI report whether the hardware is terminally wedged. For
that purposes, it only has to report the current state and not incur the
cost of checking/waiting every invocation, as we do not have to wait for
a reset before waiting on a request to ensure completion (that is baked
into the wait request implementation).

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Joonas Lahtinen <joonas.lahtinen@linux.intel.com>
Link: http://patchwork.freedesktop.org/patch/msgid/20161028125858.23563-3-chris@chris-wilson.co.uk
(cherry picked from commit c92ac094a9c647ebcff56c0535f865bc7f2aa052)
from drm-intel-next-2016-11-08
Signed-off-by: Tarun <tarun.vyas@intel.com>

BUG=b:36446508

Change-Id: I0ef71999aeede391112d93cb5cc88f4750a780c1