UPSTREAM: drm/i915: Defer setting of global seqno on request to submission
Defer the assignment of the global seqno on a request to its submission.
In the next patch, we will only allocate the global seqno at that time,
here we are just enabling the wait-for-submission before wait-for-seqno
paths.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Joonas Lahtinen <joonas.lahtinen@linux.intel.com>
Link: http://patchwork.freedesktop.org/patch/msgid/20161028125858.23563-34-chris@chris-wilson.co.uk
(cherry picked from commit f2d13290e3275df34c0cd625fbc665965af08c67)
from drm-intel-next-2016-11-08
Signed-off-by: Tarun <tarun.vyas@intel.com>

BUG=b:36446508
TEST=kmscube and relevant IGT tests along with no regressions

Change-Id: I6eaac3e7b5366a504d5aeb49342fddfa5262de87