BACKPORT: arm64: assembler: introduce ldr_this_cpu
Shortly we will want to load a percpu variable in the return from
userspace path. We can save an instruction by folding the addition of
the percpu offset into the load instruction, and this patch adds a new
helper to do so.

At the same time, we clean up this_cpu_ptr for consistency. As with
{adr,ldr,str}_l, we change the template to take the destination register
first, and name this dst. Secondly, we rename the macro to adr_this_cpu,
following the scheme of adr_l, and matching the newly added
ldr_this_cpu.

Backport Note:
- adr_this_cpu does not exist for v4.4, this commit
introduces the macro.
- The upstream change to add do_softirq_own_stack, 8e23dacd,
is not present. As a result, a subsequent changes(aa4d5d3c, d224a69e),
that used a macro to access the per-cpu to keep track of the irq_stack is
not present. As a result, the change from this_cpu_ptr to adr_this_cpu is not
required.

BUG=chromium:836330
TEST=Build and boot

Change-Id: I2ebe4ba035b45eac9b74d9e06fe198e698ee7c92
Signed-off-by: Mark Rutland <mark.rutland@arm.com>
Tested-by: Laura Abbott <labbott@redhat.com>
Cc: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Cc: James Morse <james.morse@arm.com>
Cc: Will Deacon <will.deacon@arm.com>
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>
(cherry picked from commit 1b7e2296a822dfd2349960addc42a139360ce769)
Signed-off-by: Zubin Mithra <zsm@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1037682
Reviewed-by: Guenter Roeck <groeck@chromium.org>