CHROMIUM: iio: cros_ec: Enable EC fifo int when buffer is used.
Instead of enabling the FIFO interrupt when cros_ec_ring is loaded,
enable it only when the buffer is used.
Otherwise, if the FIFO is full but the buffer is not used, the EC will
send an avalanche of interrupts.

BUG=b:35576998,chromium:785506
TEST=Check that until ARC++ is launched there are not interrupt coming
from the EC. Once ARC++ is running the interrupts are allowed.
Used " ectool motionsense fifo_int_enable".
Check that if we enable a sensor but not the buffer, no interrupts are
generated.
Check CTS sensor flush tests while enabling/disabling interrupt still pass.

Change-Id: Ia7a2cd27617c4cf19d02127533c6914824da534a
Reviewed-on: https://chromium-review.googlesource.com/412450
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
(cherry picked from commit 02934176e3952d464f5e8199250ede87935f2eaa)
Signed-off-by: Matt Atwood <matthew.s.atwood@intel.corp-partner.google.com>