CHROMIUM: ARM: dts: set phy address in mdio sub-node for tiger/fievel
Per the schematic of tiger and fievel, the phy address is 1. This patch
specify it in DT so that bus would skip scan addresses from phy.
Otherwise the RTL8211E would return an extra phyaddr 0 which is desinged
to response broadcast command and not a really phy and hence break
suspend/resume. For example if WoL is enabled, we don't want to power
down phy, while actually the extra one could power it down.

BUG=chrome-os-partner:57766
TEST=check there is only one phy(stmmac-0:01) registered.

Change-Id: Idb4564bf622b471c027b950cb8ab3bb11d641ee6
Signed-off-by: Shunqian Zheng <zhengsq@rock-chips.com>
Reviewed-on: https://chromium-review.googlesource.com/412482
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>