UPSTREAM: x86/tsc: Sync test only for the first cpu in a package
If the TSC_ADJUST MSR is available all CPUs in a package are forced to the
same value. So TSCs cannot be out of sync when the first CPU in the package
was in sync.

That allows to skip the sync test for all CPUs except the first starting
CPU in a package.

BUG=chromium:611896
TEST=do 'dmesg | grep -i tsc' on DUT, examine TSC is the clocksource,
i.e. no tsc warp.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Reviewed-by: Ingo Molnar <mingo@kernel.org>
Cc: Peter Zijlstra <peterz@infradead.org>
Cc: Yinghai Lu <yinghai@kernel.org>
Cc: Borislav Petkov <bp@alien8.de>
Link: http://lkml.kernel.org/r/20161119134017.809901363@linutronix.de
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>

(cherry picked from commit a36f5136814b6a87601220927cb9ad9ecc731e92)
Signed-off-by: Harry Pan <harry.pan@intel.com>

Change-Id: Iecc8b7e0cc8ef8f86be57889316f259f29bdd08f
Reviewed-on: https://chromium-review.googlesource.com/484245
Commit-Ready: Harry Pan <harry.pan@intel.com>
Tested-by: Harry Pan <harry.pan@intel.com>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>