UPSTREAM: backlight: pwm_bl: Move the checks for initial power state to a separate function
Move the checks to select the initial state for the backlight to a new
function and document the checks we are doing.

With the separate function it is going to be easier to fix or improve the
initial power state configuration later and it is easier to read the code.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>
Reviewed-by: Philipp Zabel <p.zabel@pengutronix.de>
Reviewed-by: Thierry Reding <treding@nvidia.com>
Signed-off-by: Lee Jones <lee.jones@linaro.org>
(cherry picked from commit 7613c922315e308a6486d802abed2eb74443dffd)

BUG=b:71057948, b:70180036
TEST=backlight remain disabled before panel drivers enable it

Change-Id: I87924461e383973a7eb1a4a8c1002f49f64da54f
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/849914
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>