UPSTREAM: HID: intel-ish-hid: Fix potential race condition
Although unlikely but it is possible that when a connect or disconnect
request is issued to the firmware, before the response comes, user
terminates the client session. In this case when the response is arrived
there is no matching client instance in the list of currently active
clients. In this case, don't issue call to wake up a waiting client.

Signed-off-by: Srinivas Pandruvada <srinivas.pandruvada@linux.intel.com>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
(cherry picked from commit 6d290391be9233349345d49f4539a87f8cb31c30)
Signed-off-by: Rushikesh S Kadam <rushikesh.s.kadam@intel.com>

BUG=b:109857156
TEST=Build kernel, boot & verify that Intel ISH driver is loaded
Verify for Trackpad input functionality (requires ISH FW to be
loaded)

Change-Id: I6e2650ffd5fed418b28c433acdaeee5744e2b4a6
Reviewed-on: https://chromium-review.googlesource.com/1096842
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Rushikesh S Kadam <rushikesh.s.kadam@intel.com>
Reviewed-by: Rushikesh S Kadam <rushikesh.s.kadam@intel.com>
Reviewed-by: Grant Grundler <grundler@chromium.org>