UPSTREAM: ext4: don't read out of bounds when checking for in-inode xattrs
With i_extra_isize equal to or close to the available space, it was
possible for us to read past the end of the inode when trying to detect
or validate in-inode xattrs.  Fix this by checking for the needed extra
space first.

This patch shouldn't have any noticeable effect on
non-corrupted/non-malicious filesystems.

BUG=chromium:766786
TEST=build/boot on caroline, xfstest

Change-Id: I50ac2d4afa0be82f6e5491bcf0eedcf97819abe7
Signed-off-by: Eric Biggers <ebiggers@google.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Reviewed-by: Andreas Dilger <adilger@dilger.ca>
(cherry picked from commit 290ab230016f187c3551d8380ea742889276d03a)
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/677692
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>