CHROMIUM: drm/udl: flush physical pages after allocating
udl_gem_get_pages gets pages which are potentially in the CPU
cache with stale contents. This memory incoherency causes
corruption which is visible on an external monitor. To fix this,
flush pages that are fetched from the cache.

This is not an issue for systems that address the cache using virtual
memory bits.

BUG=chromium:668636
TEST=trybots

Change-Id: Ie6abb4ce2fd1d0ec73fa3a4d55f8157c5d198174
Reviewed-on: https://chromium-review.googlesource.com/418180
Commit-Ready: Gurchetan Singh <gurchetansingh@chromium.org>
Tested-by: Gurchetan Singh <gurchetansingh@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>