CHROMIUM: drm/panel: simple: Change clock/timing for BOE nv101wxmn51
This change is friendly for Rockchip SoCs.

As Doug commented on CL:417662.
"We also need to make sure that this clock rate is achievable via an
integral divide of 594 MHz or 800 MHz due to the fact that we want to save
VPLL for the external display. Between 65.3 and 75 MHz, that gives;
594 / 8. = 74.25 MHz
594 / 9. = 66.00 MHz
800 / 11. = 72.7272... MHz
800 / 12. = 66.6666... MHz"

In order to support the 60 Hz refresh rate, select the 72.7272...MHZ as
the clock. This patch set the clock to 72728(MHZ), since the common clock
framework wants to round up.

BUG=chrome-os-partner:60712
TEST=boot up Bob, the panel is display.
TEST= cat sys/kernel/debug/clk/clk_summary |grep vop1
dclk_vop1_div                   1            1    72727273          0 0
dclk_vop1                    2            4    72727273          0 0
CQ-DEPEND=CL:417662, CL:414670

Original-Change-Id: Iaeb2dd5f6ca62d2c7618c6e528fea4ffab4c1ce4
Signed-off-by: Caesar Wang <wxt@rock-chips.com>
Reviewed-on: https://chromium-review.googlesource.com/418535
Reviewed-by: Douglas Anderson <dianders@chromium.org>
(cherry picked from commit fba5a6f90c2e9868aa22547b21b00a8141eb1f33)

Signed-off-by: Jeffy Chen <jeffy.chen@rock-chips.com>
Change-Id: I3bbe816a4e351891ceec62f8589710f24f6c4543
Reviewed-on: https://chromium-review.googlesource.com/444476
Commit-Queue: Caesar Wang <wxt@rock-chips.com>
Tested-by: Caesar Wang <wxt@rock-chips.com>
Reviewed-by: Douglas Anderson <dianders@chromium.org>